+++ dlib/stack/stack_kernel_1.h
@@ -97,7 +97,7 @@ namespace dlib
             ); 
 
             // functions from the enumerable interface
-            inline unsigned long size (
+            inline size_t size (
             ) const;
 
             inline bool at_start (
@@ -365,7 +365,7 @@ namespace dlib
         typename T,
         typename mem_manager
         >
-    unsigned long stack_kernel_1<T,mem_manager>::
+    size_t stack_kernel_1<T,mem_manager>::
     size (
     ) const
     {
